Updating SEP-31 title for clarity and usability
**Context**
As part of an ongoing effort to improve the usability of our protocols, in Q1 of 2021 SDF held an internal review of several SEP titles with the goal of choosing usable, descriptive names. 

Our protocols have developed over many years and as they and the ecosystem have grown, the names that they were originally given have sometimes become less meaningful. We’ve found ourselves often referring to a protocol by its number instead of its title, which can be confusing for new members as it abstracts the meaning of the SEP in discussion. While we’ll continue to use the SEP-# namespace in technical discussions and in order to maintain searchability, this step towards plain language names will hopefully encourage their usage in addition to the numeric shorthand.

Before: Direct Payments
After: Cross-border Payments API
The proposed new title of SEP-31 is “Cross-border Payments API.” The usage of "direct" didn't provide any additional clarity and could be confused with the type of direct payment performed by consumers with traditional banks.

Other names considered were Anchor-Facilitated Cross Border Payments, Anchor-Facilitated Cross-border Transfer, Mediated Payments, and Bilateral Payments. The emphasis on anchor, while clearly indicating that a payment is facilitated by anchors, is unwieldy and can be inaccurate as a party doesn’t have to be an anchor (e.g. they can be a custodial wallet with a bilateral relationship with the receiver). While “cross-border” technically at some point in the future could be a misnomer, right now it’s an important indication that the payments must be sent between two independent parties as neither has jurisdiction in the other’s country. The addition of “API” is meant to make clear the scope of the protocol.