SEP: Record dependencies in template (#913)
Record dependencies in the template of SEPs. Add a Dependencies section to the template that should be completed for new SEPs and may be added to existing SEPs if the authors of those SEPs choose.

To make it clear from first-read when there are other SEPs or CAPs that an implementer needs to be aware of and prepared to implement. In some SEPs these dependencies are buried in the text and you have to build a mental model of the dependencies you need to read an understand as you go. I think it would be convenient if these were called out up-front.