Updating SEP-1 title for clarity and usability (#919)
* Updating title for clarity and usability

As part of an ongoing effort to improve the usability of our protocols, in Q1 of 2021 SDF held an internal review of several SEP titles with the goal of choosing usable, descriptive names. 

Our protocols have developed over many years and as they and the ecosystem have grown, the names that they were originally given have sometimes become less meaningful. We’ve found ourselves often referring to a protocol by its number instead of its title, which can be confusing for new members as it abstracts the meaning of the SEP in discussion. While we’ll continue to use the SEP-# namespace in technical discussions and in order to maintain searchability, this step towards plain language names will hopefully encourage their usage in addition to the numeric shorthand.

`Stellar Info File` clearly indicates the purpose of this protocol, which was originally named simply after the file name itself.

Other names considered were Stellar Discovery File and Stellar Lookup; however, “discovery” and “lookup” don’t cover all of the information presented. 

While there was some concern over the similarity of “Info” to the `/info` endpoint, in the end we went with what felt most natural to say and was most descriptive.

* Updating readme to reflect title changes

* Updating patch version and updated date