Updating SEP-12 title for clarity and usability (#923)
* Updating SEP-12 title for clarity and usability

**Context**
As part of an ongoing effort to improve the usability of our protocols, in Q1 of 2021 SDF held an internal review of several SEP titles with the goal of choosing usable, descriptive names. 

Our protocols have developed over many years and as they and the ecosystem have grown, the names that they were originally given have sometimes become less meaningful. We’ve found ourselves often referring to a protocol by its number instead of its title, which can be confusing for new members as it abstracts the meaning of the SEP in discussion. While we’ll continue to use the SEP-# namespace in technical discussions and in order to maintain searchability, this step towards plain language names will hopefully encourage their usage in addition to the numeric shorthand.

Before: Anchor/Client customer info transfer
After: KYC API
KYC API is much easier to say than “Anchor/Client customer info transfer” and, while it’s comprised solely of acronyms, they’re both widely understood in the financial development industry. 

Spelling out Know Your Customer (KYC) in the title was considered as an educational opportunity, but it was decided that this work could be done in the body of the protocol.

* Updating patch version and updated date

* Re-updating change date and patch version