Updating SEP-24 title for clarity and usability (#924)
* Updating SEP-24 title for clarity and usability

**Context**
As part of an ongoing effort to improve the usability of our protocols, in Q1 of 2021 SDF held an internal review of several SEP titles with the goal of choosing usable, descriptive names. 

Our protocols have developed over many years and as they and the ecosystem have grown, the names that they were originally given have sometimes become less meaningful. We’ve found ourselves often referring to a protocol by its number instead of its title, which can be confusing for new members as it abstracts the meaning of the SEP in discussion. While we’ll continue to use the SEP-# namespace in technical discussions and in order to maintain searchability, this step towards plain language names will hopefully encourage their usage in addition to the numeric shorthand.

Before: Interactive Anchor/Wallet Asset Transfer Server
After: Hosted Deposit and Withdrawal
The naming considerations of SEP-24 were strongly tied to that of SEP-6, as they’re similar protocols with one major difference, that of which party hosts the collection of KYC. 

Other names considered were Interactive Asset Transfer, Webview Deposit and Withdrawal API, Anchor/Wallet Deposit API for Web, and Deposit and Withdrawal with Web UI. The usage of “on/off ramp” was noted as easier to say but less descriptive. Both “interactive” and “web ui” were considered as ways to indicate that the user would leave the client (wallet) in order to interact with the server (anchor) while submitting KYC info; “hosted” was chosen as it’s commonly used in the financial industry as “hosted checkout.” Additionally, using the words “simple” and “complex” to distinguish between the protocols was explored but found to be inaccurate as they could be seen to indicate that one protocol is more difficult to build.

A common explanation made during discussions was the comparison of SEP-6 to Stripe (a payments API) and SEP-24 to PayPal (a web UI exposed to users). More succinctly, Stripe is a Payments API and PayPal is a hosted payments page. (thanks to Jake and Kyle for introducing these metaphors). 

It should be noted that while “deposit and withdrawal” is unwieldy, there’s no perfect English word that covers those transactions in a way that meaningfully distinguishes this protocol from payments.

* Updating patch version and updated date

We discussed removing updated date completely in the future, but as that would be a change across all SEPs I'm proposing we do that later