meta: add cap diff check (#950)
Add a GitHub Action that check the diffs in CAPs are valid.

On a few occasions diffs have been committed that are corrupt. This is usually discovered by someone consuming the diff, and not by the person publishing the diff. This consumes that other persons time, as they need to debug the diff. Diffs are fiddly and it's easy for copy paste errors to occur, or stray characters, or missing new lines at the end to result in a corrupt diff. Instructions introduced in 9b0d752c7ac4d42c9735789cb190479ee8744a33 should help us generate good diffs, but this check will catch the corrupt diffs when the author is submitting them, before someone else is consuming them.

Some of the existing CAPs were easy to update to pass the check. Others not, some just don't seem to apply cleanly to any version as far as I could see, which is the type of situation this check will prevent. For those cases I added a parameter to ignore them for the moment until their authors can fix them.

This GitHub Action uses a tool, `mddiffcheck`, which lives at https://github.com/stellar/mddiffcheck. It took more effort than I expected or planned to give to it to write the tool, but it is done now.