SEP: Update Team (#990)
Remove Tom Q from the SEP team, and add Leigh and Jake.

I'm not really sure the SEP team definition is as distinct or important as the CAP team is, but it looks out of date, and myself and Jake make decisions about state transitions of SEPs. It's possible we should remove this list of people all together, but that seems like a larger conversation to have and once updated the list does seem somewhat accurate indication of the people who will typically approve SEPs.