SEP: Add details about SEP versioning (#993)
Add instructions about how to set the version number in SEPs.

We have no instructions for how to use versions, and that leads to some unfortunate situations during the draft state of a SEP. It's happened where we start a draft of a new SEP, start with v1, and then a little while later while it is still in draft we want to make a breaking change and need to change to v2. Drafts are intended to be a work-in-progress where unrestrained iteration is occurring, anything can change, and in semantic versioning that's what the v0 major version is intended for.