CAP-40: Expand on use cases in design rationale (#1022)
Expand on the use cases and impact of using CAP-40 on the number of messages.

I realized that this is not particularly clear and the design rationale did not discuss it in as much detail as we have discussed in some in person conversations. I think this fleshes it out a bit. This doesn't change the proposal in any way.