SEP-23: Rework to be a standalone specification for strkeys (#1013)
Rework SEP-23 so that it is a standalone specification for strkeys rather than a proposal to add a muxed account strkey.

SEP-23 is currently a definition of strkeys but written from the perspective of being a proposal for adding a muxed account strkey. We could really do with a single place to define all strkeys and iterate on that SEP as new strkeys are added. That is more like how other SEPs work.

We will soon propose a new strkey that will be used for CAP-40, and so I am doing this work to SEP-23 now so that it is already in the position of being the place where strkeys are specified.

This change attempts to make as least changes possible to change the intended use of the document. There are probably ways we could improve this document by adding invalid and valid test cases. We can do that, but let's do that in a different PR to this.

This change is a patch version bump because it does not introduce any breaking changes, nor introduce, change, or remove any functionality. The changes to the document are superficial and changing the intended future use of the document.