SEP-8: Distinguish requires auth vs regulated asset (#1053)
Distinguish between assets that simply require authorization from assets that are regulated assets.

An asset can require authorization without being a regulated asset or having an approval server. I think it is helpful to set this distinction and be clear that an asset requiring authorization without the issuer having an approval server is not a "SEP-8 Regulated Asset", although it may be a "regulated asset" by some other definition.