SEP-0039: Changing NFT metadata's `url` field to use an HTTP gateway
After some discussion within the Stellar Quest community, relating to a
newly released side quest all about NFTs, we noticed the SEP-0039 example
metadata uses the `ipfs://` protocol scheme for the metadata field of `url`.
It seems like it would be more compatible with existing ecosystem practices
to use an HTTP(S) gateway in this field, rather than the direct IPFS url.

This PR would implement exactly that suggestion in the example JSON within
this document. I've used the cloudflare URL that is already used later on in
the document, under the 'Representing NFTs' section.

Signed-off-by: Elliot Voris <elliot@voris.me>