CAP-72 - Switch to the approach of modifying the AccountEntry directly
After describing the contract data-based approach in the first iteration, it became pretty apparent that this approach is more complex than I've originally anticipated, and just extending the `Signer` seems like a much easier way to go (as proven by the -100 lines of diffs in this PR). Also, if we want to do trustline management from Soroban, we'd need to do pretty much the same thing anyways.